+++ src/main/app/src/app/formulieren/taken/model/aanvullende-informatie-task-form.ts
@@ -11,7 +11,7 @@ import { lastValueFrom } from "rxjs";
 import { InformatieObjectenService } from "../../../informatie-objecten/informatie-objecten.service";
 import { KlantenService } from "../../../klanten/klanten.service";
 import { MailtemplateService } from "../../../mailtemplate/mailtemplate.service";
-import { FormField } from "../../../shared/form/dynamic-form/dynamic-form";
+import { FormField } from "../../../shared/form/composed-form/composed-form.component";
 import { GeneratedType } from "../../../shared/utils/generated-types";
 import { ZakenService } from "../../../zaken/zaken.service";
 import { OptionValue } from "../taak.utils";
