+++ src/main/app/src/app/formulieren/taken/model/abstract-task-form.spec.ts
@@ -24,7 +24,7 @@ import { RouterModule } from "@angular/router";
 import { TranslateModule } from "@ngx-translate/core";
 import { of } from "rxjs";
 import { fromPartial } from "../../../../test-helpers";
-import { FormField, ZacDynamicForm as ZacForm } from "../../../shared/form/dynamic-form/dynamic-form";
+import { FormField, ZacComposedForm } from "../../../shared/form/composed-form/composed-form.component";
 import { MaterialFormBuilderModule } from "../../../shared/material-form-builder/material-form-builder.module";
 import { GeneratedType } from "../../../shared/utils/generated-types";
 import { AbstractTaskForm } from "./abstract-task-form";
@@ -116,10 +116,10 @@ class TestForm extends AbstractTaskForm {
 describe(AbstractTaskForm.name, () => {
   let formulier: TestForm;
   let fixture: ComponentFixture<
-    ZacForm<Record<string, AbstractControl<unknown, unknown>>>
+    ZacComposedForm<Record<string, AbstractControl<unknown, unknown>>>
   >;
   let componentRef: ComponentRef<
-    ZacForm<Record<string, AbstractControl<unknown, unknown>>>
+    ZacComposedForm<Record<string, AbstractControl<unknown, unknown>>>
   >;
   let loader: HarnessLoader;
   let formGroup: FormGroup;
@@ -153,7 +153,7 @@ describe(AbstractTaskForm.name, () => {
         );
       }
 
-      fixture = TestBed.createComponent(ZacForm);
+      fixture = TestBed.createComponent(ZacComposedForm);
       componentRef = fixture.componentRef;
       componentRef.setInput("fields", fields);
       componentRef.setInput("form", formGroup);
