+++ src/main/app/src/app/formulieren/taken/model/document-verzenden-post-task-form.ts
@@ -9,7 +9,7 @@ import moment from "moment";
 import { lastValueFrom } from "rxjs";
 import { mapStringToDocumentenStrings } from "../../../documenten/document-utils";
 import { InformatieObjectenService } from "../../../informatie-objecten/informatie-objecten.service";
-import { FormField } from "../../../shared/form/dynamic-form/dynamic-form";
+import { FormField } from "../../../shared/form/composed-form/composed-form.component";
 import { GeneratedType } from "../../../shared/utils/generated-types";
 import { AbstractTaskForm } from "./abstract-task-form";
