+++ src/main/app/src/app/plan-items/human-task-do/human-task-do.component.html
@@ -23,7 +23,7 @@
 
 <!-- form using Angular form builder -->
 <section *ngIf="formFields.length" class="mt-2">
-  <zac-dynamic-form
+  <zac-composed-form
     [form]="form"
     [fields]="formFields"
     [config]="_formConfig"
