+++ src/main/app/src/app/taken/taak-view/taak-view.component.html
@@ -65,7 +65,7 @@
     <div class="flex-row flex-wrap gap-10">
       <mat-card class="taak-formulier w50 flex">
         <mat-card-content>
-          <zac-dynamic-form
+          <zac-composed-form
             *ngIf="formFields.length"
             [form]="form"
             [fields]="formFields"
